requirements.txt
pandas==1.2.3
matplotlib==3.3.4
torch==1.8.0
tqdm==4.58.0
torchvision==0.9.0
numpy==1.19.2
dataset==1.4.5
Pillow==8.1.2
scikit_learn==0.24.1
